src/image_indexer/requirements.txt
pypdf
langchain
langchain-text-splitters
langchain-core
langchain-community
langchain-openai
python-dotenv
azure-cognitiveservices-speech
azure-servicebus==7.13.0
azure-cosmos==4.8.0
azure-storage-blob==12.24.0